archive/migration/requirements.txt
# Garden6 MongoDB to Supabase Migration Dependencies
psycopg2-binary>=2.9.0
pyyaml>=6.0